+++ packages/apps/session-storage/shopify-app-session-storage-dynamodb/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @shopify/shopify-app-session-storage-dynamodb
 
+## 4.0.5
+
 ## 4.0.4
 
 ### Patch Changes